Major changes to code base
Small changes include fixing of typos, changing variable names
and descriptions of functions.

Larger changes include completing the class structure of the
electrongas.py and the dielectric_class.py files. Additionally, there
are modifications to the __init__.py files to improve the
importing of submodules.

This new structure also changes the focus of this project to be on the
dielectric function, and leave the dynamic structure factor and stopping
powers to serve as applications of this project.

"Do one thing and do it well"